+++ spec/rubocop/cops/lint/empty_ensure_spec.rb
@@ -1,0 +1,33 @@
+# encoding: utf-8
+
+require 'spec_helper'
+
+module Rubocop
+  module Cop
+    module Lint
+      describe EmptyEnsure do
+        let(:cop) { EmptyEnsure.new }
+
+        it 'registers an offence for empty ensure' do
+          inspect_source(cop,
+                         ['begin',
+                          '  something',
+                          'ensure',
+                          'end'])
+          expect(cop.offences).to have(1).item
+        end
+
+        it 'does not register an offence for non-empty ensure' do
+          inspect_source(cop,
+                         ['begin',
+                          '  something',
+                          '  return',
+                          'ensure',
+                          '  file.close',
+                          'end'])
+          expect(cop.offences).to be_empty
+        end
+      end
+    end
+  end
+end